requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
cffi==1.17.1
cryptography==43.0.1
frozenlist==1.4.1
idna==3.10
importlib_metadata==8.5.0
multidict==6.1.0
platformdirs==4.3.3
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
tomli==2.0.1
typing_extensions==4.12.2
yapf==0.40.2
yarl==1.11.1
zipp==3.20.2
